p.feature: Finally settled on using the library LuaSQL.
While it doesn't provide the best abstractions nor does it provide good LSP support, it was
the least troublesome to implement and make work. To ease the troubles, I'm writing some
wrapper functions around common SQL operations that will make its usage more straightforward.

These wrappers are not supposed to be a full-featured solution, for now it'll be just the things
this project needs to work properly. I might in the future expand this wrapper and release it
separately.